backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pyyaml==6.0.2
treys==0.1.8
